# [IMP] Code cleanup and speed improvement reading XML/HTML fields