The old algorythm was recreating a full etree from scratch, while parsing the content to translate. Now, we just parse the etree and update only the content that to be translated.

Before:
| In [2]: %timeit views.refresh(); views.mapped('arch_db')
| 504 ms ± 18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

After:
| In [2]: %timeit views.refresh(); views.mapped('arch_db')
| 191 ms ± 2.38 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Reading a view is 2.63x faster: from 504 ms, to 191 ms (sum of all the 1345 views, when installing website_sale). Large views, like web pages goes up to 3x-4x faster; small views are ~2x faster.

As views are cached, it only impacts the loading to put in cache, but all XML/HTML fields get the same performance gain.


--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr