@fpodoo 2021-03-23
> Sorry to insist, but I am pretty sure in this case the outcome will be the same.
My bad; you are right. I learnt something today :)
We will apply your improvement during the merge.